Seperate out swiss map marshaling to its own file
